Refactor io types
