+++ src/storage/logging.rs
@@ -8,8 +8,9 @@ use std::ops::Deref;
 use std::sync::{Arc, Mutex, Weak};
 
 use crate::errors::{Error, ErrorKind, Result};
-use crate::helpers::{writable_or_create, Deferred};
-use crate::io::{IOEvent, IdType, InputType};
+use crate::helpers::{Deferred, writable_or_create};
+use crate::io::{IdType, IOEvent};
+use crate::io::InputType;
 use crate::settings::Settings;
 use crate::storage::{Container, MappedCollection, Persistent};
